move case,overlay models to core package
