+++ src/app/packages/core/models/index.ts
@@ -1,0 +1,6 @@
+export { Overlay, OverlaysCriteria } from './overlay.model';
+export { Position } from './position.model';
+export { MenuItem } from './menu-item.model';
+export { IGeneralConfig } from './general.config';
+export { Case } from './case.model';
+export { Case, CaseState, CaseTimeState, CaseFacetsState, CaseMapsState, CaseMapState, defaultMapType } from './case.model';
